Current state. Still not working
