+++ packages/itwinui-react/src/index.ts
@@ -3,6 +3,6 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 export * from './core';
-export * from './types/base-react-table';
+// export * from './types/base-react-table';
 export * from './types/react-table-config';
 // export * from './types/base-react-table.d.ts';
